requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
python-pptx
scikit-learn
nltk
numpy